feat: add support for NUCLEO-F446RE board
- Add NUCLEO-F446RE environment to platformio.ini
- Use mbed upload protocol for the new board
- Fix type mismatch error in cmdRead for STM32 compatibility

+++ firmware/src/cmd.cpp
@@ -183,7 +183,7 @@ static void cmdRead(command_session *session) {
             return;
         }
 
-        len = min(len, max_rom_size - addr);
+        len = min((int)len, (int)(max_rom_size - addr));
         p = strtok(0, TOK_SEP);
     }
